File Tree
This is the version of the project that will open up hidden until you
click on the tags. The concept is a div hiding function, unhide. This
function is an action that is performed onclick of a button and acts on
a div with a certain id. We may have to edit the CSS to make it look
prettier.
--Pranav