Flow Database
Hey yall I know its been some time
I added the flow database, it should work, From now on, flows should be
saved with a name, the content and tags. However, the tags can just be
an empty string.
The search method for flows is now just getCard, it returns a card based
on the name.
All the other methods are relatively the same with different names,
instead of File**** it will be changed to Flow****.
if you want to reference the flow database, call db2, if you want to
reference the regular database, call db